Feedback

+++ hw02/Problem1.java
@@ -4,14 +4,14 @@ public class Problem1{
  
         public static void main(String arg []){
 
-        int width;
-        int height;
+        double width;
+        double height;
 
         Scanner Kb = new Scanner(System.in);
 
         System.out.println("Please enter the width and height of a rectangle: ");
-        width = Kb.nextInt();
-        height = Kb.nextInt();
+        width = Kb.nextDouble();
+        height = Kb.nextDouble();
         
         System.out.println("The perimeter is " + ((2*width)+(2*height)));
         System.out.println("The Area is "+(width*height));
